+++ pcs/stonith.py
@@ -148,7 +148,8 @@ def stonith_list_options(lib, argv, modifiers):
 
     print(resource._format_agent_description(
         lib.stonith_agent.describe_agent(agent_name),
-        True
+        stonith=True,
+        show_advanced=modifiers["full"]
     ))
 
 def stonith_create(lib, argv, modifiers):